Added class GOMidiPort https://github.com/GrandOrgue/grandorgue/issues/703
